feat: Add reusable OtherProject component to replace inline logic
Replaced inline code for "OtherProject" functionality with a new, reusable `OtherProject` component. This simplifies the structure of `ResultProjectList.vue` and `CatalogProjects.vue`, reducing redundancy and improving maintainability. Unneeded logic and imports have been removed accordingly.
